style: add controller menu panel and clean HUD pattern
